requirements-dev.txt
# Development dependencies for testing
Pillow>=10.0.0